requirements.txt
#可用pip freeze > ./requirements.txt 生成此文件(依赖包以及环境)
#pip install -r requirements.txt    可以通过该命令进行安装
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
coverage==4.5.4
cryptography==2.9.2
Django==2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
filelock==3.0.12
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
idna==2.8
importlib-metadata==0.20
Jinja2==2.11.2
jsonpatch==1.24
jsonpointer==2.0
lxml==4.5.2
MarkupSafe==1.1.1
more-itertools==7.2.0
mysqlclient @ https://download.lfd.uci.edu/pythonlibs/w3jqiv8s/mysqlclient-1.4.6-cp36-cp36m-win32.whl
nose2==0.9.1
numpy==1.17.1
packaging==19.1
paramiko==2.7.1
pdfkit==0.6.1
pexpect==4.7.0
Pillow==6.1.0
pipenv==2018.11.26
pluggy==0.12.0
ptyprocess==0.6.0
py==1.8.0
pycparser==2.20
Pygments==2.4.2
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.2
PyPubSub==3.3.0
pyserial==3.4
pytest==5.1.2
pytz==2019.3
pywin32==224
requests==2.22.0
robotframework==3.1.2
robotframework-autoitlibrary==1.2.4
robotframework-httplibrary==0.4.2
robotframework-requests==0.5.0
robotframework-ride==1.7.3.1
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==3.3.1
robotframeworklexer==1.1
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
toml==0.10.0
toolz==0.10.0
tox==3.14.0
urllib3==1.25.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
waitress==1.3.1
wcwidth==0.1.7
WebOb==1.8.5
WebTest==2.0.33
wkhtmltopdf==0.2
wxPython==4.0.6
zipp==0.6.0
zope.event==4.4
zope.interface==5.1.0
